Enable test on CI and add new template
