+++ eng/testing/scenarios/BuildWasiAppsJobsList.txt
@@ -3,3 +3,4 @@ Wasi.Build.Tests.ILStripTests
 Wasi.Build.Tests.SdkMissingTests
 Wasi.Build.Tests.RuntimeConfigTests
 Wasi.Build.Tests.WasiTemplateTests
+Wasi.Build.Tests.PInvokeTableGeneratorTests
