+++ src/mono/wasm/templates/templates/wasi-native/local.c
