+++ src/mono/wasm/templates/templates/wasi-native/runtimeconfig.template.json
@@ -1,0 +1,10 @@
+{
+    "wasmHostProperties": {
+        "perHostConfig": [
+            {
+                "name": "wasmtime",
+                "Host": "wasmtime"
+            }
+        ]
+    }
+}
